requirements.txt
aniso8601==3.0.2
cachetools==2.1.0
Click==7.0
Flask==1.0.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
httplib2==0.11.3
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysql-connector==2.1.6
oauth2client==4.1.3
parse==1.8.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
pytz==2018.5
rsa==4.0
six==1.11.0
SQLAlchemy==1.2.12
uritemplate==3.0.0
Werkzeug==0.14.1
